feat(node): add child_process.fork
