+++ node/_tools/config.json
@@ -22,6 +22,7 @@
       "test-buffer-indexof.js",
       "test-dns-lookup.js",
       "test-dns.js",
+      "test-child-process-fork.js",
       "test-event-emitter-errors.js",
       "test-event-emitter-max-listeners.js",
       "test-event-emitter-no-error-provided-to-error-event.js",
@@ -106,6 +107,7 @@
       "test-buffer-zero-fill-cli.js",
       "test-buffer-zero-fill-reset.js",
       "test-buffer-zero-fill.js",
+      "test-child-process-fork.js",
       "test-dns-lookup.js",
       "test-dns.js",
       "test-event-emitter-add-listeners.js",
