+++ node/_tools/suites/parallel/test-child-process-fork.js
@@ -1,0 +1,72 @@
+// deno-fmt-ignore-file
+// deno-lint-ignore-file
+
+// Copyright Joyent and Node contributors. All rights reserved. MIT license.
+// Taken from Node 16.13.0
+// This file is automatically generated by "node/_tools/setup.ts". Do not modify this file manually
+
+// Copyright Joyent, Inc. and other Node contributors.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a
+// copy of this software and associated documentation files (the
+// "Software"), to deal in the Software without restriction, including
+// without limitation the rights to use, copy, modify, merge, publish,
+// distribute, sublicense, and/or sell copies of the Software, and to permit
+// persons to whom the Software is furnished to do so, subject to the
+// following conditions:
+//
+// The above copyright notice and this permission notice shall be included
+// in all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
+// OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN
+// NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+// DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
+// OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE
+// USE OR OTHER DEALINGS IN THE SOFTWARE.
+// Flags: --no-warnings
+'use strict';
+const common = require('../common');
+const assert = require('assert');
+const { fork } = require('child_process');
+const args = ['foo', 'bar'];
+const fixtures = require('../common/fixtures');
+const debug = require('util').debuglog('test');
+
+const n = fork(fixtures.path('child-process-spawn-node.js'), args);
+
+assert.strictEqual(n.channel, n._channel);
+assert.deepStrictEqual(args, ['foo', 'bar']);
+
+n.on('message', (m) => {
+  debug('PARENT got message:', m);
+  assert.ok(m.foo);
+});
+
+// https://github.com/joyent/node/issues/2355 - JSON.stringify(undefined)
+// returns "undefined" but JSON.parse() cannot parse that...
+/* TODO(kt3k): Uncomment the below test cases when the ipc features are ready.
+assert.throws(() => n.send(undefined), {
+  name: 'TypeError',
+  message: 'The "message" argument must be specified',
+  code: 'ERR_MISSING_ARGS'
+});
+assert.throws(() => n.send(), {
+  name: 'TypeError',
+  message: 'The "message" argument must be specified',
+  code: 'ERR_MISSING_ARGS'
+});
+
+assert.throws(() => n.send(Symbol()), {
+  name: 'TypeError',
+  message: 'The "message" argument must be one of type string,' +
+           ' object, number, or boolean. Received type symbol (Symbol())',
+  code: 'ERR_INVALID_ARG_TYPE'
+});
+n.send({ hello: 'world' });
+*/
+
+n.on('exit', common.mustCall((c) => {
+  assert.strictEqual(c, 0);
+}));
